requirements.txt
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
curl_cffi==0.12.0
debugpy==1.8.15
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
findspark==2.0.1
frozendict==2.4.6
greenlet==3.2.3
h11==0.16.0
hvac==2.3.0
idna==3.10
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kafka-python==2.2.15
matplotlib-inline==0.1.7
multitasking==0.0.12
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
parso==0.8.4
peewee==3.18.2
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py4j==0.10.9.9
pycparser==2.22
Pygments==2.19.2
PySocks==1.7.1
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pyzmq==27.0.0
requests==2.32.4
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.42
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yfinance==0.2.65